Address yaml errors in multus playbook
Using ansible version 2.7.2 on Fedora 29-server, I'm unable to run the
multus provisioning pb due to errors in the quoting of os commands.

This patch just converts those items to use double quotes instead of
single quotes, which I believe is also compatable with older (and newer)
versions of ansible.